pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tiger_audit_trail"
version = "0.1.0"
description = "Enterprise audit trail for Django with diffing and async delivery"
readme = "README.md"
requires-python = ">=3.13"
authors = [{name = "AJAY", email = "ajay@tigerlab.com"}]
license = {text = "MIT"}
dependencies = [
    "Django>=5.2",
    "djangorestframework>=3.14",
    "celery>=5.3",
    "boto3>=1.28",
    "confluent-kafka>=2.0",
    "pymongo>=4.4",
    "httpx>=0.24",
    "django-cors-headers>=4.1",
    "nltk>=3.9",
    "cryptography>=43.0",
]
keywords = ["audit", "django", "compliance", "logging"]

[project.optional-dependencies]
dev = ["pytest", "pytest-django", "moto", "ruff", "black", "sphinx", "furo", "sphinx-autodoc-typehints", "myst-parser", "sphinx-copybutton", "sphinx-rtd-theme", "sphinx-autobuild"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["audit_trail*"]
